combined_cibil_bank_django/requirements.txt
asgiref
asttokens
backcall
certifi
charset-normalizer
click
cloudpickle
comm
dask
debugpy
decorator
distributed
Django
django-cors-headers
django-lrucache-backend
django-material
django-pandas
django-picklefield
django-request-logging
djangorestframework
exceptiongroup
executing
featuretools
fsspec
idna
importlib-metadata
ipykernel
ipython
jedi
Jinja2
joblib
jupyter_client
jupyter_core
locket
MarkupSafe
matplotlib-inline
msgpack
nest-asyncio
numpy
packaging
pandas
parso
partd
pexpect
pickleshare
pip
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
pyarrow
Pygments
python-dateutil
pytz
PyYAML
pyzmq
requests
scikit-learn
scipy
setuptools
six
sortedcontainers
sqlparse
stack-data
tblib
threadpoolctl
toolz
tornado
tqdm
traitlets
typing_extensions
urllib3
wcwidth
wheel
zict
zipp
